Drop support for PHP 5.4
