+++ composer.json
@@ -10,7 +10,7 @@
     "license": "BSD-3-Clause",
     "type": "library",
     "require": {
-        "php": ">=5.4.0",
+        "php": ">=5.5.0",
         "symfony/config": "^2.3.0|^3|^4|^5|^6.0|^7.0",
         "symfony/dependency-injection": "^2.3.0|^3|^4|^5|^6.0|^7.0",
         "symfony/filesystem": "^2.3.0|^3|^4|^5|^6.0|^7.0",
